Add files via upload
Data record for storage in csv. It has an interface to choose the type of sausage, working hours, type of machine and record the data and notes about the machine to keep a record of the maintenance and operation of the machines in the meat laboratory of the National University of Colombia, Medellín headquarters.